Add CI/CD workflow to PRs and main pushes
- Configure verification job to lint, type-check, and build
- Configure deploy job to connect to VPS and build changes
- Fix linting errors in web
