+++ apps/web/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@pishposh/web",
   "private": true,
-  "version": "0.5.2",
+  "version": "0.5.3",
   "type": "module",
   "scripts": {
     "dev": "vite",
