+++ apps/web/src/pages/CreatePost.tsx
@@ -1,6 +1,6 @@
 import { useState, type SubmitEvent } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext'
+import { useAuth } from '../hooks/useAuth'
 import { createPost } from '../services/posts'
 
 export function CreatePost() {
