+++ apps/web/src/pages/Login.tsx
@@ -1,6 +1,6 @@
 import { useState, type SubmitEvent } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext'
+import { useAuth } from '../hooks/useAuth'
 
 export function Login() {
   const [email, setEmail] = useState('')
