+++ apps/web/src/pages/Signup.tsx
@@ -1,7 +1,7 @@
 import { Turnstile, type TurnstileInstance } from '@marsidev/react-turnstile'
 import { useRef, useState, type SubmitEvent } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext'
+import { useAuth } from '../hooks/useAuth'
 import { ApiError } from '../services/auth'
 
 export function Signup() {
